+++ Linux/driver/umd/src/device/simulator/umemory.h
@@ -126,15 +126,15 @@ public:
     aipu_status_t free_all(void);
     virtual bool invalid(uint64_t addr) const;
     virtual bool get_info(uint64_t addr, uint64_t &base, uint32_t &size) const;
-    virtual int read(uint64_t addr, void *dest, size_t size) const
+    virtual int64_t read(uint64_t addr, void *dest, size_t size) const
     {
         return mem_read(addr, dest, size);
     };
-    virtual int write(uint64_t addr, const void *src, size_t size)
+    virtual int64_t write(uint64_t addr, const void *src, size_t size)
     {
         return mem_write(addr, src, size);
     };
-    virtual int zeroize(uint64_t addr, size_t size)
+    virtual int64_t zeroize(uint64_t addr, size_t size)
     {
         return mem_bzero(addr, size);
     };
